jic/mysite/requirements.txt
Django==5.2
wagtail==7.0
django-livereload-server==0.4
django-compressor==4.5.1
minio==7.2.3
Pillow>=10.2.0,<12.0.0
